refactor(oracle-adapter): consolidate AdminChanged and FeederAdded into events.rs
